Fix endless loop bug for handling PHP errors
Previous version used readlines() to handle error stream when
starting phpsh; however, if an error is present in phpsh.php then readlines()
waits endlessly for additional errors. Because of this, exceptions are
not raised to the user and program does not terminate. Bug was fixed
by utilizing readline() and only outputting a
single bug at a time (which is likely fine, since bugs could be fixed
one at a time).